REF: move join tests
